requirements.txt
python-telegram-bot[job-queue]==20.7
psycopg2-binary==2.9.9
yt-dlp
youtube-transcript-api
groq
instaloader
pdfplumber==0.11.0
google-cloud-vision==3.7.2
anthropic>=0.40.0
gspread
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
